+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+- `cli.get_aws_credentials()` passes `id_token` to `GetId` / `GetCredentialsForIdentity` Logins instead of `access_token`.
+- All Cognito keys try env first and fall back to `~/.toshi/auth_config.json` file.
+
+### Added
+- `NZSHM22_TOSHI_COGNITO_IDENTITY_POOL_ID` env var so that `~/.toshi/auth_config.json` is not required.
+
 ## [1.2.3] - 2026-05-22
 
 ### Added